Let CommandLineRunnerWPF inherit from CommandLineRunner.
