+++ src/DynamoCLI/CommandLineRunner.cs
@@ -55,7 +55,6 @@ namespace DynamoCLI
             var stateNames = new List<String>();
             stateNames.Add("default");
 
-            var outputresults = new List<Dictionary<Guid, List<object>>>();
             XmlDocument doc = null;
             foreach (var stateName in stateNames)
             {
@@ -70,33 +69,9 @@ namespace DynamoCLI
                 //if verbose was true, then print all nodes to the console
                 if (!String.IsNullOrEmpty(cmdLineArgs.Verbose))
                 {
-                    doc = new XmlDocument();
-                    var resultsdict = new Dictionary<Guid, List<object>>();
-                    foreach (var node in model.CurrentWorkspace.Nodes)
-                    {
-                        var portvalues = new List<object>();
-                        foreach (var port in node.OutPorts)
-                        {
-                            var value = node.GetValue(port.Index, model.EngineController);
-                            if (value.IsCollection)
-                            {
-                                portvalues.Add(GetStringRepOfCollection(value));
-                            }
-                            else if (value.IsDictionary)
-                            {
-                                portvalues.Add(GetStringRepOfDictionary(value.Data));
-                            }
-                            else
-                            {
-                                portvalues.Add(value.StringData);
-                            }
-                        }
-
-                        resultsdict.Add(node.GUID, portvalues);
-                    }
-                    outputresults.Add(resultsdict);
-                    populateXmlDocWithResults(doc, outputresults);
+                    doc = CreateXMLDoc(model);
                 }
+
                 evalComplete = false;
 
             }
@@ -104,7 +79,7 @@ namespace DynamoCLI
             return doc;
         }
 
-        private static string GetStringRepOfCollection(ProtoCore.Mirror.MirrorData value)
+        protected static string GetStringRepOfCollection(ProtoCore.Mirror.MirrorData value)
         {
             var items = string.Join(",",
                 value.GetElements().Select(x =>
@@ -115,7 +90,7 @@ namespace DynamoCLI
             return "{" + items + "}";
         }
 
-        private static string GetStringRepOfDictionary(object value)
+        protected static string GetStringRepOfDictionary(object value)
         {
             if (value is DesignScript.Builtin.Dictionary || value is IDictionary)
             {
@@ -173,6 +148,40 @@ namespace DynamoCLI
                 }
 
             }
+        }
+
+        protected static XmlDocument CreateXMLDoc(DynamoModel model)
+        {
+            var outputresults = new List<Dictionary<Guid, List<object>>>();
+            XmlDocument doc = new XmlDocument();
+            var resultsdict = new Dictionary<Guid, List<object>>();
+
+            foreach (var node in model.CurrentWorkspace.Nodes)
+            {
+                var portvalues = new List<object>();
+                foreach (var port in node.OutPorts)
+                {
+                    var value = node.GetValue(port.Index, model.EngineController);
+                    if (value.IsCollection)
+                    {
+                        portvalues.Add(GetStringRepOfCollection(value));
+                    }
+                    else if (value.IsDictionary)
+                    {
+                        portvalues.Add(GetStringRepOfDictionary(value.Data));
+                    }
+                    else
+                    {
+                        portvalues.Add(value.StringData);
+                    }
+                }
+
+                resultsdict.Add(node.GUID, portvalues);
+            }
+            outputresults.Add(resultsdict);
+            populateXmlDocWithResults(doc, outputresults);
+
+            return doc;
         }
 
         private static void OpenWorkspaceAndConvert(DynamoModel model, string dynPath)
